backend/geocoding-function/requirements.txt
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.18.1
pandas==2.1.4
functions-framework==3.5.0
requests==2.31.0
db-dtypes==1.2.0